+++ OpenDiscordHaxx/Raidbot/Bot.cs
@@ -2,6 +2,7 @@
 {
     public abstract class Bot
     {
+        public int Threads { get; protected set; }
         public bool ShouldStop { get; set; }
         public Attack Attack { get; protected set; }