+++ OpenDiscordHaxx/Raidbot/Leaver.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 using Discord;
 
@@ -14,14 +13,17 @@ namespace DiscordHaxx
         {
             Attack = new Attack(this) { Type = RaidOpcode.Leave, Bots = Server.Bots.Count };
 
-
+            Threads = request.Threads;
             _guildId = request.GuildId;
+
+            if (_guildId <= 0)
+                throw new CheckException("Invalid guild ID");
         }
 
 
         public override void Start()
         {
-            Parallel.ForEach(new List<DiscordClient>(Server.Bots), new ParallelOptions() { MaxDegreeOfParallelism = 2 }, bot =>
+            Parallel.ForEach(new List<DiscordClient>(Server.Bots), new ParallelOptions() { MaxDegreeOfParallelism = Threads }, bot =>
             {
                 try
                 {