+++ OpenDiscordHaxx/Socket/Bot/Requests/ModResponse.cs
@@ -1,0 +1,18 @@
+﻿using Newtonsoft.Json;
+
+namespace DiscordHaxx
+{
+    class ModResponse : BotRequest
+    {
+        public ModResponse() : base(BotOpcode.BotModify)
+        { }
+
+
+        [JsonProperty("at")]
+        public string At { get; set; }
+
+
+        [JsonProperty("success")]
+        public bool Success { get; set; }
+    }
+}