+++ Changelog.md
@@ -1,3 +1,12 @@
+# 0.0.2.1
+
+- Added support for setting custom thread amounts on attacks.
+- Added Threads to the dashboard's 'Ongoing attacks' bot list.
+- Added more input handling.
+- Fixed a CPU usage issue.
+
+
+
 # 0.0.2.0 (rewrite)
 
 ### Additions